Move functions out of uvdata for better reusability
